Add subdirectory for all x402 tools.
