--- typescript/lib/x402-mcp-tool/src/index.ts
@@ -1,1 +1,0 @@
-export { xServerTool } from "./server.js";
