+++ typescript/lib/x402/mcp-tool/package.json
