+++ typescript/lib/x402/mcp-tool/src/index.ts
@@ -6,11 +6,11 @@ import type {
   PaymentRequirements,
   Price,
   Resource,
+  VerifyResponse,
 } from "x402/types";
 import { useFacilitator } from "x402/verify";
 import { processPriceToAtomicAmount } from "x402/shared";
 import z, { type ZodRawShape } from "zod";
-import type { X402PaymentResponse } from "./client.js";
 import type {
   McpServer,
   RegisteredTool,
@@ -29,6 +29,16 @@ if (!facilitatorUrl || !payTo) {
 
 const { verify, settle } = useFacilitator({ url: facilitatorUrl });
 const x402Version = 1;
+
+/**
+ * The response from the X402 payment when there is an error
+ */
+export interface X402PaymentResponse {
+  x402Version: number;
+  error?: string;
+  payer?: VerifyResponse["payer"];
+  accepts: PaymentRequirements[];
+}
 
 /**
  * Creates payment requirements for a given price and network
