+++ typescript/lib/x402/mcp-tool/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     "moduleResolution": "Node16",
     "outDir": "dist",
